Refactor

+++ internal/model/entity/product.go
@@ -1,7 +1,7 @@
 package entity
 
 type Product struct {
-	ID    int64  `json:"-" gorm:"primary_key;AUTO_INCREMENT"`
+	ID    uint   `json:"-" gorm:"primary_key;AUTO_INCREMENT"`
 	Name  string `json:"name"`
 	Price int    `json:"price"`
 }
